requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
cloudinary==1.24.0
Django==3.1.3
django-admin-rangefilter==0.6.4
django-bootstrap-modal-forms==2.0.1
django-ckeditor==6.0.0
django-countries==6.1.3
django-crispy-forms==1.10.0
django-csp==3.7
django-extensions==3.1.0
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
django-timezone-field==4.1.1
dnspython==2.0.0
gunicorn==20.0.4
idna==2.10
phonenumbers==8.12.13
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.0
twilio==6.48.0
urllib3==1.26.2
whitenoise==5.2.0
